# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vessl API Documentation",
        "title": "Vessl API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/projects": {
            "get": {
                "description": "ListProjects endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vercel"
                ],
                "summary": "ListProjects endpoint",
                "responses": {}
            },
            "post": {
                "description": "CreateProject endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "CreateProject endpoint",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vessl_dev_vessl_internal_models.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/projects/{id}": {
            "get": {
                "description": "GetProject endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "GetProject endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "DeleteProject endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "DeleteProject endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "vessl_dev_vessl_internal_models.CreateProjectRequest": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "internalPort": {
                    "type": "integer"
                },
                "internal_port": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "repositoryUrl": {
                    "type": "string"
                },
                "repository_url": {
                    "type": "string"
                },
                "teamId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}